# artifacts/afucloud-backend/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AfuCloud API – cloud platform for storing, managing, and delivering digital assets
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RefreshInput {
  refreshToken: string;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  token: string;
  /** @minLength 8 */
  password: string;
}

export interface ProfileUpdate {
  name?: string;
  /** @nullable */
  avatar?: string | null;
}

export interface User {
  id: string;
  email: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  emailVerified?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  accessToken: string;
  refreshToken: string;
}

export interface Project {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  userId: string;
  imageCount?: number;
  storageUsed?: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ProjectInput {
  name: string;
  description?: string;
}

export interface ProjectUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
}

export interface ProjectStats {
  totalImages: number;
  storageUsed: number;
  bandwidth: number;
  apiRequests: number;
  deletedImages?: number;
  favoriteImages?: number;
}

export interface Image {
  id: string;
  projectId: string;
  name: string;
  /** @nullable */
  originalName?: string | null;
  url: string;
  /** @nullable */
  publicUrl?: string | null;
  downloadUrl?: string;
  format: string;
  size: number;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  favorite?: boolean;
  tags?: string[];
  /** @nullable */
  album?: string | null;
  /** @nullable */
  deletedAt?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ImageListResponse {
  images: Image[];
  total: number;
  page: number;
  limit: number;
}

export interface ImageUpdate {
  name?: string;
  tags?: string[];
  /** @nullable */
  album?: string | null;
}

export interface UploadUrlInput {
  filename: string;
  contentType: string;
  name?: string;
}

export interface UploadUrlResponse {
  uploadUrl: string;
  imageId: string;
  key: string;
}

export interface ConfirmUploadInput {
  imageId: string;
  key: string;
  width?: number;
  height?: number;
  size?: number;
  tags?: string[];
  album?: string;
}

export interface ApiKey {
  id: string;
  projectId: string;
  name: string;
  environment: string;
  prefix: string;
  scopes?: string[];
  /** @nullable */
  lastUsedAt?: string | null;
  createdAt: string;
}

export interface ApiKeyInput {
  name: string;
  environment: string;
  scopes?: string[];
}

export type ApiKeyWithSecret = ApiKey & {
  secret: string;
};

export interface Token {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  prefix: string;
  scopes?: string[];
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  lastUsedAt?: string | null;
  /** @nullable */
  revokedAt?: string | null;
  createdAt: string;
}

export interface TokenInput {
  name: string;
  description?: string;
  scopes?: string[];
  /** @nullable */
  expiresAt?: string | null;
}

export type TokenWithSecret = Token & {
  secret: string;
};

export interface AnalyticsOverview {
  totalProjects: number;
  totalImages: number;
  totalStorageUsed: number;
  totalBandwidth: number;
  totalApiRequests: number;
  recentUploads?: number;
}

export interface DailyStat {
  date: string;
  uploads: number;
  downloads: number;
  bandwidth: number;
}

export interface ProjectAnalytics {
  uploads: number;
  downloads: number;
  storageUsed: number;
  bandwidth: number;
  apiRequests: number;
  dailyStats: DailyStat[];
}

export interface Webhook {
  id: string;
  projectId: string;
  url: string;
  events: string[];
  active: boolean;
  createdAt: string;
}

export interface WebhookInput {
  url: string;
  events: string[];
  secret?: string;
}

export interface WebhookUpdate {
  url?: string;
  events?: string[];
  active?: boolean;
  /** @nullable */
  secret?: string | null;
}

export type ActivityLogMetadata = { [key: string]: unknown };

export interface ActivityLog {
  id: string;
  userId: string;
  /** @nullable */
  projectId?: string | null;
  action: string;
  resource: string;
  /** @nullable */
  resourceId?: string | null;
  metadata?: ActivityLogMetadata;
  createdAt: string;
}

export type ListImagesParams = {
search?: string;
tag?: string;
album?: string;
format?: string;
favorite?: string;
page?: string;
limit?: string;
};

export type GetProjectAnalyticsParams = {
period?: string;
};

export type ListActivityParams = {
projectId?: string;
limit?: string;
};