# artifacts/afucloud-backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: AfuCloud API – cloud platform for storing, managing, and delivering digital assets
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /v1/auth/me/update:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /v1/auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /v1/auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── PROJECTS ───
...
  /v1/projects/{projectId}/api-keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List API keys for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Create a new API key
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: API key created (secret shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyWithSecret"
  /v1/projects/{projectId}/api-keys/{id}:
    delete:
      operationId: revokeApiKey
      tags: [api-keys]
      summary: Revoke an API key
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── PERSONAL ACCESS TOKENS ───
...
  /v1/projects/{projectId}/analytics:
    get:
      operationId: getProjectAnalytics
      tags: [analytics]
      summary: Analytics for a specific project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Project analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAnalytics"

  # ─── WEBHOOKS ───
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── Auth ───
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: ["string", "null"]
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        avatar:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user, accessToken, refreshToken]
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
        refreshToken:
          type: string

    # ─── Projects ───
...
    ApiKey:
      type: object
      required: [id, projectId, name, environment, prefix, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        environment:
          type: string
        prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        lastUsedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ApiKeyInput:
      type: object
      required: [name, environment]
      properties:
        name:
          type: string
        environment:
          type: string
        scopes:
          type: array
          items:
            type: string
    ApiKeyWithSecret:
      allOf:
        - $ref: "#/components/schemas/ApiKey"
        - type: object
          required: [secret]
          properties:
            secret:
              type: string

    # ─── Tokens ───
...
    ProjectAnalytics:
      type: object
      required: [uploads, downloads, storageUsed, bandwidth, apiRequests, dailyStats]
      properties:
        uploads:
          type: number
        downloads:
          type: number
        storageUsed:
          type: number
        bandwidth:
          type: number
        apiRequests:
          type: number
        dailyStats:
          type: array
          items:
            $ref: "#/components/schemas/DailyStat"
    DailyStat:
      type: object
      required: [date, uploads, downloads, bandwidth]
      properties:
        date:
          type: string
        uploads:
          type: number
        downloads:
          type: number
        bandwidth:
          type: number

    # ─── Webhooks ───
...